patch: extract RidgeSettings to its own module
